Fixes a bug when extracting integers larger than 32-bits due to the signatures using only a 32-bit result. Catch any exceptions thrown during string extraction. Leaker.mem_* will now choose the correct 32-bit or 64-bit call based on the .bits property instead of depending on how the control itself is compiled.
disassembler:
    Changed all bits and width properties to an unsigned size_t.
    Changed the calculation of the width for dumping floating-point
        numbers in scientific notation to use a string for length.

Leaker:
    Changed all instances of LONG for the bits attribute and
        dumping/disassembling length to an unsigned ULONG.
    When extracting a string-type from the address-space, catch
        the exception if an access-violation occurs.
    Promoted the 32-bit values for extracting integers to a proper
        64-bit type to avoid losing the top 32-bits.
    Modified the VirtualQueryEx wrappers to use the correct API
        based on the number of bits configured in the disassembler.

+++ Ax/disassembler.h
@@ -14,7 +14,7 @@ protected:
 
 public:
 	/* public properties */
-	int m_bits;
+	size_t m_bits;
 	enum cs_opt_value m_syntax;
 
 public:
@@ -64,7 +64,7 @@ public:
 
 	enum cs_opt_value syntax(enum cs_opt_value);
 
-	int bits(int num);
+	size_t bits(size_t num);
 
 	size_t size(intptr_t ea, size_t count);
 	size_t disasm(intptr_t ea, size_t count, std::ostream& os);
@@ -77,8 +77,8 @@ public:
 
 private:
 	/* private members */
-	int m_bits;
-	int m_width;
+	size_t m_bits;
+	size_t m_width;
 
 	const char unprintable = '.';
 	const std::string divider = " | ";
@@ -128,7 +128,7 @@ protected:
 
 public:
 	/* public interface */
-	Dumper(int bits, int width) : m_bits(bits), m_width(width) {}
+	Dumper(size_t bits, size_t width) : m_bits(bits), m_width(width) {}
 	~Dumper() {}
 
 	template <typename T>
